[GraphQL] Remove "Live" ObjectKind, rename "Historical" to "Indexed" (#17313)
## Description

With the removal of direct accesses to the `objects` table, the
nomenclature we use around object kinds ("not indexed", "live",
"historical", "deleted or wrapped") is less relevant. In particular, we
never use "live" and the name "historical" is confusing because even the
latest version of the object may be considered "historical".

This PR removes the "live" kind, and renames "historical" to "indexed"
to better reflect how the users of GraphQL perceive the kinds of
objects, rather than how the data is stored in the indexer's tables.

## Test plan

```
sui$ cargo nextest run -p sui-graphql-rpc
sui$ cargo nextest run -p sui-graphql-e2e-tests --features pg_integration
```

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [x] GraphQL: There is no longer a `Live` or `Historical`,
`ObjectKind`, they have been merged into a single `Indexed` kind
representing any object that has been fetched from the index (as opposed
to an object that we have information about but that hasn't been indexed
yet, or an object that has been deleted or wrapped so is not available).
- [ ] CLI: 
- [ ] Rust SDK: