Add quantity_min_tick to DeepBook. Quantity must be divisible by this amount, rather than lot_size. (#17335)
## Description 

We have some pools with a large lot_size. All trade quantities must be a
multiple of lot_size, so they must be truncated to meet this
requirement. Assets being bridged from another chain utilize all decimal
points, and truncating these decimals results in users losing non
negligible amount of assets. For example: USDC has 6 decimal points, and
a user may bridge 10.129218 USDC. If we attempt to swap that amount into
SUI, the user's USDC is truncated to 10.00000.

QUANTITY_MIN_TICK is a constant that's set to 1000, this will make it so
only the last 3 decimal points are truncated. In the above example, the
user will be able to swap 10.129000 USDC, and their loss amount is
negligible.

Essentially, lot_size now becomes min_size, and quantity_min_tick is now
lot_size. We can't rename lot_size because that would change the API.
This is not as bad as it sounds.

Unit tests were affected heavily because input / output amounts were now
validated by 1000. Unit tests were originally written with a lot_size of
1, which in my opinion is not reflective of a real world scenario. I had
to 10x a lot of the values.

## Test plan 

Unit tests. Code changes are not invasive.

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [x] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------

Co-authored-by: Aslan Tashtanov <aslantashtanov@Aslans-MacBook-Pro.local>