Update test util for creating shared objects (#17582)
## Description 

We have two test utils that create a random shared object
- shared_for_testing, which returns a shared object always with the same
ID when calling from the same thread
- with_id_shared_for_testing, which returns a random shared object each
time called

Looking at the history of shared_for_testing, it was first created to be
used in a specific test, where a random object ID
is hard coded. Later on, the function is gradually moved to be a global
test util, but the move always tries to preserve the
function semantic that when called multiple times, same object ID is
used. Particularly, test_certificates() expects using the same shared
object as the tests calling this function.

Therefore, we probably don't want to have a shared_for_testing with a
surprising semantics. This PR merges
with_id_shared_for_testing into shared_for_testing so that it always
returns a random shared object, and explicitly
pass the shared object when calling test_certificates()

## Test plan 

Unit test

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: