[consensus] Hookup LeaderSchedule change & Recovery paths (#17040)
## Description 
This is one of many PRs to implement Leader Scoring & Leader Schedule
feature for Mysticeti.
- [Design
Doc](https://www.notion.so/mystenlabs/Leader-Schedule-Scoring-538c65cb370b4644944a94485efc4979?pvs=4)
- [Prototype PR](https://github.com/MystenLabs/sui/pull/16975)

Leader Schedule Change Notes:
- We can choose the scoring strategy via an env var similar to how we
are doing for networking. Will submit the workflow changes so this is
easily configurable via gh deploy workflows
- In practice it looks like the LeaderSchedule change takes about
[250](https://metrics.sui.io/goto/5pPPAExSR?orgId=1)-[500ms](https://metrics.sui.io/goto/qifU0ExIR?orgId=1)
depending on the scoring strategy chosen. While this is not negatively
impacting performance it does mean that there is additional room for
improvement in a follow up PR with how we are traversing the DAG

Recovery Notes:
- We are now only writing `last_commit_rounds` & `reputation_scores` on
a leader schedule change. Any commits that happen after that will be
accounted for during recovery.

![Screenshot 2024-04-03 at 10 07
44 PM](https://github.com/MystenLabs/sui/assets/97870774/b1ec55da-13c1-43a6-b3fe-e3c3eeff8843)
![Screenshot 2024-04-03 at 10 07
12 PM](https://github.com/MystenLabs/sui/assets/97870774/bf2804fc-0787-4e8d-a5c4-2e2c855264cb)

## Results
![Screenshot 2024-04-03 at 10 23
41 PM](https://github.com/MystenLabs/sui/assets/97870774/7e14a58d-d455-41f3-8b9f-aeb7142c8fb0)


## Test Plan 

unit tests