[indexer] properly set key for cache in indexer reader (#17649)
## Description 

Right now we are setting the key of the cache to be the raw string `"{
format!("{}{}", package_id, obj_type) }"`, which does not change with
the input causing the cache to essentially store the first entry and
always returning it. This PR fixes that. I've scanned the code base and
changed the only two occurrences of `cache_get_or_set_with`

## Test plan 

tested locally

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [x] Indexer: 
- [x] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: