[cleanup][move] Remove stackless bytecode interpreter (#17763)
## Description 

The stackless bytecode interpreter was only being used in unit tests,
and was not being used there at all. The only usages were for usages to
test itself. Since it's no longer of any real use/not being used it's
best to remove the code to remove the ongoing maintenance burden of it.

This removes the stackless bytecode interpreter from the unit tests, and
then remove the stackless bytecode interpreter crate and
bytecode-testsuite crate for it as well.

## Test plan 

Make sure existing tests pass.