[0/n][enums] Move enums implementation (#17245)
## Description 

This adds the enums implementation for Move. This PR only contains the
changes for enums within `external-crates/move` and nowhere else. This
PR _does not_ contain any tests. The tests for these changes can be
found in the PR above in the stack.

Note that individual commits are cut out within this PR to (hopefully)
make the review process a bit easier.

## Stack:
* #17245 **<< You are here**
* #17246 
* #17247 
* #17248 
* #17249 
* #17250 
* #17251 

## Test plan 

Tested in the PR above this. 



---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [X] Protocol: Adds a new protocol version, and enables Move enums on
devnet.
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [X] JSON-RPC: Adds support for Move enum values in returned json-rpc
results.
- [X] GraphQL: Adds support for Move enum values and types to GraphQL.
- [ ] CLI: 
- [ ] Rust SDK: