[move] Fix and add interactive disassembler (#17986)
## Description 

Fixes the interactive disassembler, and adds the flag for it to `sui
move disassemble` (with the `-i` flag). See screencast for what this
looks like.



https://github.com/MystenLabs/sui/assets/2895723/50a75eb1-3d84-4c00-a30b-ab051c2ca169



## Test plan 

👀 

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [X] CLI: Adds new `-i` flag to `sui move disassemble` that creates an
interactive disassembled bytecode-to-source explorer.
- [ ] Rust SDK: