Initial metrics work for the indexer (#18025)
## Description 

Initial work for monitoring values. 
This is a good start for values to track and please advise on what else
you'd like to see.

## Test plan 

Not much now and need to come up with something to check that things are
good

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: