[consensus] Disable multi leader per round in universal committer (#18206)
## Description 

On the boundaries of Leader Schedule change it is possible to elect a
new leader for the same round resulting in multiple commits for the same
round. This was unintended output from consensus as we have set the
parameters to only commit 1 leader per round. Sui also only expects one
commit per round, because of this the second commit of the same round is
ignored by consensus handler. This led to an error in simtests where the
`EndOfPublish` transaction was part of this ignored commit and we could
not form a quorum to close the epoch without this node.

Changes
- Explicitly set number of leaders per round to 1 for Mysticeti commits.
I wanted to try and maintain this but it would be a much more intrusive
change and we will probably have to do a larger refactor to fully
support it anyways, so better off doing the cleaner mitigation for now.
- Pass back the block that the transaction was included in as part of
the tx acknowledgment and include that in the logging to help with
debugging future issues.
- Also a minor change to print if all tests pass during simtest
seed-search

## Test plan 

```
❯ scripts/simtest/seed-search.py simtest --test test_reconfig_with_committee_change_basic --num-seeds 1000
    Updating git repository `https://github.com/MystenLabs/mysten-sim.git`
   Compiling consensus-core v0.1.0 (/Users/arunkoshy/Documents/GitHub/sui/consensus/core)
   ...
   All tests passed successfully!
  Killing child processes
  [1]    391 killed     scripts/simtest/seed-search.py simtest --test  --num-seeds 1000
```

---

## Release notes

- [X] Protocol: Version 50 - Explicitly set number of leaders per round
to 1 for Mysticeti commits
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK:

---------

Co-authored-by: Mark Logan <mark@mystenlabs.com>