[move-ide] Added support for enums (#18124)
## Description 

This PR adds enums support to the Move analyzer. 

In this PR, I also sneaked a fix for symbols merging, which was not very
robust when multiple packages are built within the same IDE instance.
It's much more robust to keep symbols on the per-package basis, even if
it's more memory-intensive.

## Test plan 

All existing and newly added tests must pass.