Prevent effects equivocation (#18426)
With the writeback cache enabled, a validator can lose execution outputs
when restarting, if they have not yet been committed to disk. But the
validator may have returned signed effects for those transactions to
users. If they have done so, we must be sure we do not produce differing
outputs when re-executing any such transaction. (The most likely way
this would happen would be due to switching to an upgraded binary that
changes execution in some way).