[move-ide] Added tests for inlay type hints (#18484)
## Description 

What the title says. I decided to go for by-hand tests rather than add a
framework similar to what we use for symbols as we really only need a
rather limited number of tests and they will be different for different
types of inlay hints (e.g. for upcoming parameter hints) so by-hand
tests seem to be more cost-efficient.

## Test plan 

All new and existing tests must pass