move: manage [addresses] automatically on publish / upgrade (#18470)
## Description 

This PR enables automation of package addresses so that users don't need
to fiddle with the `[addresses]` section when publishing/upgrading a
package. However, if they are accustomed to the previous habits of
changing `[addresses]`, they can still do so: the PR is backwards
compatible with the "old way" described in the docs, which we'll keep
during transition. In that case, though, they can't benefit from
publishing to multiple chains.

## Test plan 

Existing tests + updated upgrade happy path test to exercise Part 1 and
Part 2 above + local testing against localnet, testnet

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: Running `sui client publish` and `sui client upgrade` will now manage addresses when publishing a package to one or more chains. Refer to the documentation on Automated Address Management for usage and guidelines.
- [ ] Rust SDK: