[cli] Deprecate sui-test-validator (#18515)
## Description 

Deprecate sui-test-validator but keep the binary with a migration
message. Since moving all workflows to use `sui start` this week, all
seems to be working as expected, thus we can proceed with deprecating
the logic in this crate.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [x] CLI: Starting with v1.29, `sui-test-validator` binary will no
longer be used to start a network. If you used it in the past, download
the latest release archive and use the `sui-pg` binary if you need to
start an indexer or GraphQL service, or build from source with the
indexer feature enabled: `cargo build --bin sui --features indexer`.
You can also use the `sui/scripts/sui-test-validator.sh` drop-in
replacement script to run the commands as you used to.

If you only used `sui-test-validator` to start a local network from a
new genesis, use `sui start --with-faucet --force-regenesis` which will
exhibit the exact same behaviour.
Finally, if you only used `sui start` before, the behaviour is preserved
and you can continue using it as before.

- [ ] Rust SDK:
---------

Co-authored-by: Ashok Menon <amenon94@gmail.com>