rest: create and host openapi spec and explorer (#18529)
## Description 

This PR introduces an OpenAPI spec for the REST api and hosts the spec
and an explorer off of the main rest service. While the spec as
introduced in this PR enumerates all of the available endpoints, the
documentation, inputs, request and response types for all of the
endpoints has not been completed and will be done in a follow up PR. The
spec can be explored locally (without the actual backend service) by
running:

```
OPENAPI_EXPLORER=1 cargo test -p sui-rest-api openapi_explorer
```

and navigating to `http://127.0.0.1:8000/openapi` in a browser.

The openapi spec is checked-in to the repo and a test exists to ensure
that the checked-in version matches what is derived by the service. This
allows for subsequent changes to the spec to be reviewed at code review
time.

The api spec is still under development and is not to be considered
stable yet.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to
your changes, release notes aren't required.

For each box you select, include information after the relevant heading
that describes the impact of your changes that a user might notice and
any actions they must take to implement updates.

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: